pyproject.toml
[tool.poetry]
name = "brilliantimagery-ui"
version = "0.3.0"
description = ""
authors = ["Chad DeRosier <chad.derosier@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.1"
pillow = "^7.0.0"
scikit-video = "^1.1.11"
numpy = "^1.18.1"
pyside2 = "^5.14.1"
brilliantimagery = "^0.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
auto-py-to-exe = "^2.6.6"
pywin32-ctypes = "^0.2.0"
pefile = "^2019.4.18"
pywin32 = "^227"
pytest-qt = "^3.3.0"
pytest-cov = "^2.8.1"
pytest-mock = "^2.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
